Create siswa controller for siswa
